ml-service/requirements.txt
# Deep Learning Framework (PyTorch CPU)
torch>=2.5.0

# Core ML Libraries (Python 3.13 compatible)
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.3.0

# API Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
flask>=3.0.0
flask-cors>=4.0.0

# Geospatial Utilities
geopy>=2.4.1

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
python-multipart>=0.0.6
requests>=2.31.0

# Development dependencies
pytest>=7.4.3
black>=23.12.0
flake8>=7.0.0
